FEAT:             // Evita que o arquivo de mesmo nome anterior seja deletado, deleta arquivo antigo

+++ app/Services/SobreMimService.php
@@ -21,13 +21,23 @@ class SobreMimService
             $extensao = $curriculo->getClientOriginalExtension();
             $nomeArquivo = $nomeOriginal . '_' . now()->format('d-m-Y') . '.' . $extensao;
             $urlCurriculo = $curriculo->storeAs('sobre-mim', $nomeArquivo, 'public');
-            $deleteCurriculo = Storage::disk('public')->delete($sobreMim->ds_url_curriculo);
+            // Evita que o arquivo de mesmo nome anterior seja deletado, deleta arquivo antigo
+            if($sobreMim->ds_url_curriculo != $urlCurriculo){
+                $deleteCurriculo = Storage::disk('public')->delete($sobreMim->ds_url_curriculo);
+            }
         }
 
         if($request->file('ds_url_foto_usuario')){
             $foto = $request->file('ds_url_foto_usuario');
-            $urlFoto = $foto->store('sobre-mim', 'public');
-            $deleteFoto = Storage::disk('public')->delete($sobreMim->ds_url_foto_usuario);
+            $nomeOriginal = pathinfo($foto->getClientOriginalName(), PATHINFO_FILENAME);
+            $extensao = $foto->getClientOriginalExtension();
+            $nomeArquivo = $nomeOriginal . '.' . $extensao;
+            $urlFoto = $foto->storeAs('sobre-mim', $nomeArquivo, 'public');
+            // Evita que o arquivo de mesmo nome anterior seja deletado, deleta arquivo antigo
+            if($sobreMim->ds_url_foto_usuario != $urlFoto){
+                $deleteFoto = Storage::disk('public')->delete($sobreMim->ds_url_foto_usuario);
+            }
+
         }
 
         $countDeletes = 0;
